backend/pyproject.toml
[tool.poetry]
name = "participation-api"
version = "0.1.0"
description = ""
authors = ["Rodrigo Saraiva <rodrigobezerrasaraiva@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.7"
djangorestframework = "^3.14.0"
django-extensions = "^3.2.1"
gunicorn = "^20.1.0"
psycopg2-binary = "^2.9.5"
pytest = "^7.2.2"
pytest-django = "^4.5.2"
factory-boy = "^3.2.1"
django-filter = "^22.1"
dj-rest-auth = {extras = ["with-social"], version = "^3.0.0"}
djangorestframework-simplejwt = "^5.2.2"

[tool.poetry.group.dev.dependencies]
blue = "^0.9.1"
isort = "^5.12.0"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "setup.settings"
python_files = "tests.py test_*.py *_tests.py"
testpaths = [
    "tests",
]
filterwarnings = "ignore::DeprecationWarning:pkg_resources.*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
